requirements.txt
Appium-Python-Client>=5.2.0
selenium>=4.0.0
pytest>=8.0.0
pytest-html>=4.0.0
pytest-xdist>=3.0.0
requests>=2.28.0
websocket-client>=1.6.0
